Nice refactoring 2015-06 to reduce duplication
