+++ crates/ornament/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 comfy-table = { version = "7.1.4", features = ["custom_styling"] }
+nom.workspace = true
 owo-colors = "4.2.0"
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde_json.workspace = true
