+++ crates/ornament/src/parsing.rs
@@ -1,0 +1,1 @@
+pub type ParseError<'a> = nom::Err<nom::error::Error<&'a str>>;
