+++ puzzles/2015/2015-06/src/lib.rs
@@ -1,3 +1,4 @@
+mod grid;
 mod instruction;
 mod part1;
 mod part2;
