+++ puzzles/2015/2015-06/src/part2.rs
@@ -1,32 +1,22 @@
-use crate::instruction::{Instruction, InstructionKind};
+use crate::{
+    grid::Grid,
+    instruction::{Instruction, Task},
+};
 
 pub fn part_2(input: &str) -> usize {
-    input
-        .trim()
-        .lines()
-        .map(|line| line.trim().parse::<Instruction>().unwrap())
-        .fold(vec![[0usize; 1000]; 1000], |mut grid, instruction| {
-            for column in grid
-                .iter_mut()
-                .take(instruction.to.x + 1)
-                .skip(instruction.from.x)
-            {
-                for light in column
-                    .iter_mut()
-                    .take(instruction.to.y + 1)
-                    .skip(instruction.from.y)
-                {
-                    match instruction.kind {
-                        InstructionKind::Toggle => *light += 2,
-                        InstructionKind::Enable => *light += 1,
-                        InstructionKind::Disable => *light = light.saturating_sub(1),
-                    }
-                }
-            }
+    let mut grid = Grid::new(1000, 0usize);
 
-            grid
-        })
-        .iter()
-        .flat_map(|column| column.iter())
-        .sum()
+    grid.process(
+        input
+            .trim()
+            .lines()
+            .map(|line| Instruction::parse(line).unwrap()),
+        |task, light| match task {
+            Task::Toggle => *light += 2,
+            Task::Enable => *light += 1,
+            Task::Disable => *light = light.saturating_sub(1),
+        },
+    );
+
+    grid.columns().flat_map(|column| column.iter()).sum()
 }
